agent/requirements-window.txt
aiofiles                                 25.1.0
aiohappyeyeballs                         2.6.1
aiohttp                                  3.13.2
aiosignal                                1.4.0
annotated-types                          0.7.0
anyio                                    4.12.0
attrs                                    25.4.0
audioread                                3.1.0
av                                       16.0.1
certifi                                  2025.11.12
cffi                                     2.0.0
charset-normalizer                       3.4.4
click                                    8.1.8
colorama                                 0.4.6
coloredlogs                              15.0.1
comtypes                                 1.4.14
decorator                                5.2.1
distro                                   1.9.0
docstring_parser                         0.17.0
eval_type_backport                       0.3.1
filelock                                 3.20.0
flatbuffers                              25.9.23
frozenlist                               1.8.0
fsspec                                   2025.10.0
googleapis-common-protos                 1.72.0
grpcio                                   1.76.0
gTTS                                     2.5.4
h11                                      0.16.0
httpcore                                 1.0.9
httpx                                    0.28.1
huggingface-hub                          0.36.0
humanfriendly                            10.0
idna                                     3.11
importlib_metadata                       8.7.0
iniconfig                                2.3.0
Jinja2                                   3.1.6
jiter                                    0.12.0
joblib                                   1.5.3
lazy_loader                              0.4
librosa                                  0.11.0
livekit                                  1.0.23
livekit-agents                           1.3.10
livekit-api                              1.1.0
livekit-blingfire                        1.0.0
livekit-plugins-cartesia                 1.3.6
livekit-plugins-deepgram                 1.3.6
livekit-plugins-elevenlabs               1.3.6
livekit-plugins-noise-cancellation       0.2.5
livekit-plugins-openai                   1.3.6
livekit-plugins-silero                   1.3.5
livekit-plugins-turn-detector            1.3.5
livekit-protocol                         1.1.1
llvmlite                                 0.46.0
markdown-it-py                           4.0.0
MarkupSafe                               3.0.3
mdurl                                    0.1.2
mpmath                                   1.3.0
msgpack                                  1.1.2
multidict                                6.7.0
nest-asyncio                             1.6.0
networkx                                 3.5
numba                                    0.63.1
numpy                                    2.3.5
ollama                                   0.6.1
onnxruntime                              1.23.1
openai                                   2.8.1
opentelemetry-api                        1.39.0
opentelemetry-exporter-otlp              1.39.0
opentelemetry-exporter-otlp-proto-common 1.39.0
opentelemetry-exporter-otlp-proto-grpc   1.39.0
opentelemetry-exporter-otlp-proto-http   1.39.0
opentelemetry-proto                      1.39.0
opentelemetry-sdk                        1.39.0
opentelemetry-semantic-conventions       0.60b0
packaging                                25.0
pillow                                   11.3.0
pip                                      25.3
piper-tts                                1.3.0
platformdirs                             4.5.1
pluggy                                   1.6.0
pooch                                    1.8.2
prometheus_client                        0.23.1
propcache                                0.4.1
protobuf                                 6.33.1
psutil                                   7.1.3
PyAudio                                  0.2.14
pycparser                                2.23
pydantic                                 2.12.5
pydantic_core                            2.41.5
pydub                                    0.25.1
Pygments                                 2.19.2
PyJWT                                    2.10.1
pypiwin32                                223
pyreadline3                              3.5.4
pytest                                   9.0.1
pytest-asyncio                           1.3.0
python-dotenv                            1.2.1
pyttsx3                                  2.99
pywin32                                  311
PyYAML                                   6.0.3
regex                                    2025.11.3
requests                                 2.32.5
rich                                     14.2.0
ruff                                     0.14.7
safetensors                              0.7.0
scikit-learn                             1.8.0
scipy                                    1.17.0
setuptools                               70.2.0
shellingham                              1.5.4
sniffio                                  1.3.1
sounddevice                              0.5.3
soundfile                                0.13.1
soxr                                     1.0.0
srt                                      3.5.3
sympy                                    1.14.0
threadpoolctl                            3.6.0
tokenizers                               0.22.1
torch                                    2.9.1+cpu
torchaudio                               2.9.1+cpu
torchvision                              0.24.1+cpu
tqdm                                     4.67.1
transformers                             4.57.1
typer                                    0.20.0
types-protobuf                           6.32.1.20251105
typing_extensions                        4.15.0
typing-inspection                        0.4.2
urllib3                                  2.5.0
vosk                                     0.3.45
watchfiles                               1.1.1
websockets                               15.0.1
yarl                                     1.22.0
zipp                                     3.23.0
